backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.22.0
pydantic-settings>=2.2.1
sqlalchemy>=2.0.25
alembic>=1.13.1
chromadb==0.4.24
sentence-transformers>=2.2.2
httpx>=0.24.1
python-multipart>=0.0.6
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
mysql-connector-python>=8.0.33
numpy<2.0
scikit-learn>=1.0
pdfplumber>=0.10
python-docx>=1.1.0
python-pptx>=0.6.23
pandas>=2.0
joblib>=1.3
datasets>=2.19
transformers>=4.40
torch>=2.2
accelerate>=0.29

